Fix boundary and UB issues across small recognizers/enumerators
- simple_quadrangulation_enum: drop the bogus even-n restriction (odd
  n exist: Herschel graph at n=11); cite OEIS A007022 with correct
  values instead of a mistranscribed A078666; check_simple_quadrangulation
  early-exit comment now matches the 3-connected definition (cube, n=8)
- ktree: edgeless graphs are 0-trees per the recursive definition, so
  check_ktree accepts them and enumerate(0,0) returns the empty graph,
  making recognizer and enumerator consistent for k=0
- poset/chain/cograph enumerators: n=0 returns the empty structure
  (A001035(0)=1; matches the other enumerators and the recognizers)
- circle.h: initialize CircleResult::is_circle (was uninitialized UB)
- tournament: check_tournament(0, arcs) now rejects stray arcs,
  consistent with check_digraph
- interval (AT_FREE): set is_interval only after the clique-path
  certificate is built, so unreachable failure paths fail closed
  instead of returning YES with empty intervals
- laman_enum: widen subset bitmask shifts to 64-bit (32-bit shift UB)
- digraph/poset CLI: clamp negative n/m like read_directed instead of
  aborting on reserve(-1)

+++ include/laman_enum.h
@@ -77,11 +77,11 @@ inline bool is_23_sparse(const LamanEnumState& state, int x) {
     // Only check subsets containing vertex x
     // sub is a bitmask representing a subset of {1,...,x-1} (sub >= 1)
     unsigned int x_minus_1 = static_cast<unsigned int>(x - 1);
-    for (unsigned int sub = 1; sub < (1u << x_minus_1); ++sub) {
+    for (unsigned long long sub = 1; sub < (1ULL << x_minus_1); ++sub) {
         // k = |sub| + 1 (including vertex x)
         int k_minus_1 = 0;
         {
-            unsigned int tmp = sub;
+            unsigned long long tmp = sub;
             while (tmp) { tmp &= (tmp - 1); ++k_minus_1; }
         }
         int k = k_minus_1 + 1;
@@ -93,7 +93,7 @@ inline bool is_23_sparse(const LamanEnumState& state, int x) {
 
         // Edges from vertex x
         for (int i = 0; i < static_cast<int>(x_minus_1) && !exceeded; ++i) {
-            if (!(sub & (1u << i))) continue;
+            if (!(sub & (1ULL << i))) continue;
             if (state.adj[x][i + 1]) {
                 ++edges;
                 if (edges > limit) exceeded = true;
@@ -102,10 +102,10 @@ inline bool is_23_sparse(const LamanEnumState& state, int x) {
 
         // Edges within {1,...,x-1}
         for (int i = 0; i < static_cast<int>(x_minus_1) && !exceeded; ++i) {
-            if (!(sub & (1u << i))) continue;
+            if (!(sub & (1ULL << i))) continue;
             for (int j = i + 1; j < static_cast<int>(x_minus_1) && !exceeded;
                  ++j) {
-                if (!(sub & (1u << j))) continue;
+                if (!(sub & (1ULL << j))) continue;
                 if (state.adj[i + 1][j + 1]) {
                     ++edges;
                     if (edges > limit) exceeded = true;
